rewrite with mypy
