+++ ambisonic_decode/type_definitions.py
@@ -1,0 +1,2 @@
+Rad = float  # angles in radian
+
