+++ requirements.txt
@@ -1,12 +1,20 @@
 attrs==19.3.0
+cycler==0.10.0
 importlib-metadata==1.6.0
+kiwisolver==1.2.0
+matplotlib==3.2.1
 more-itertools==8.2.0
+mypy==0.770
+mypy-extensions==0.4.3
 numpy==1.18.2
 packaging==20.3
 pluggy==0.13.1
 py==1.8.1
 pyparsing==2.4.7
 pytest==5.4.1
+python-dateutil==2.8.1
 six==1.14.0
+typed-ast==1.4.1
+typing-extensions==3.7.4.2
 wcwidth==0.1.9
 zipp==3.1.0
